handle exceptions and errores

+++ src/main/java/com/tickers/io/applicationapi/services/AuthenticationService.java
@@ -1,14 +1,11 @@
 package com.tickers.io.applicationapi.services;
 import com.tickers.io.applicationapi.exceptions.BadRequestException;
-import com.tickers.io.applicationapi.exceptions.NotFoundException;
 import com.tickers.io.applicationapi.exceptions.UnauthenticatedException;
-import com.tickers.io.applicationapi.jwt.KeyProvider;
 import com.tickers.io.applicationapi.jwt.KeyWithId;
 import com.tickers.io.applicationapi.model.User;
 import com.tickers.io.applicationapi.model.UserSession;
 import com.tickers.io.applicationapi.repositories.UserRepository;
 import com.tickers.io.applicationapi.repositories.UserSessionsRepository;
-//import com.tickers.io.applicationapi.support.TenantContext;
 import com.tickers.io.applicationapi.utils.RequestUtils;
 import io.jsonwebtoken.*;
 import com.tickers.io.applicationapi.utils.TransactionHandler;
@@ -20,28 +17,32 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.Builder;
 import lombok.Getter;
-import org.checkerframework.checker.units.qual.K;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
 
-import java.nio.charset.StandardCharsets;
 import java.security.Key;
 import java.time.Instant;
 import java.time.ZonedDateTime;
 import java.time.temporal.ChronoUnit;
-import java.util.Arrays;
-import java.util.Date;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 public class AuthenticationService {
+    private final String HOST_DOMAIN = "api.tickers.local";
+
+    private final String ISSUER = "api.mlservice.local";
+
+    private final String SECRET = "Rdg0gZsfsubWzajsbT8xPVnmd1BASyXojy9/PAV5+VrE5OVq+e/7xA==";
+
     @Autowired
     private HttpServletResponse response;
+
+    @Autowired
+    private HttpServletRequest request;
 
     @Autowired
     private PasswordEncoder passwordEncoder;
@@ -59,15 +60,12 @@ public class AuthenticationService {
 
     private Logger logger = LoggerFactory.getLogger(AuthenticationService.class);
 
-    private String secret = "Rdg0gZsfsubWzajsbT8xPVnmd1BASyXojy9/PAV5+VrE5OVq+e/7xA==";
-
     @PostConstruct
     protected void init() {
-//        SigningKeyResolver x509SigningKeyResolver = new X509SigningKeyResolver();
         SigningKeyResolver key = new SigningKeyResolverAdapter(); // TODO: refactor here
         this.jwtParser = Jwts.parserBuilder()
-                .requireAudience("api.tickers.local")
-                .requireIssuer("api.mlservice.local")
+                .requireAudience(this.HOST_DOMAIN)
+                .requireIssuer(this.ISSUER)
                 .setSigningKeyResolver(new SigningKeyResolverAdapter() {
                     @Override
                     public byte[] resolveSigningKeyBytes(JwsHeader header, Claims claims) {
@@ -96,10 +94,12 @@ public class AuthenticationService {
     }
 
     public void authenticate(AuthenticationContext context) {
-        User user = userRepository.findOneByUserName(context.getUsername()).orElseThrow(NotFoundException::new);
+        User user = userRepository.findOneByUserName(context.getUsername()).orElse(null);
+        if (user == null) {
+            throw UnauthenticatedException.INVALID_CREDENTIALS;
+        }
 
         if (user.getBlocked()) throw new BadRequestException();
-//        tenantContext.setTenantId(UUID.fromString("7d8d66ab-e025-420b-bc63-2845b6319b99"));
 
         transactionHandler.runInNewTransaction(() -> {
             // Check the credentials, if not a SAML auth request
@@ -112,22 +112,35 @@ public class AuthenticationService {
             String accessToken = this.generateAccessToken(context, user, session);
             String refreshToken = this.generateRefreshToken(context, user, session);
 
+           this.removeAllCookieHaveDomain(request, response);
+
             //Set the cookies
             this.injectAuthenticationTokenCookie(response, accessToken);
             this.injectRefreshTokenCookie(response, refreshToken);
         });
     }
 
-    public String resolveToken(HttpServletRequest httpServletRequest) {
-        String bearerToken = httpServletRequest.getHeader("Authorization");
-        if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7);
-        }
-        return null;
-    }
-
     public Jws<Claims> parseJwt(String token) {
         return this.jwtParser.parseClaimsJws(token);
+    }
+
+    public Jws<Claims> parseCookieFilterJwt(String token) throws
+            UnsupportedJwtException,
+            MalformedJwtException,
+            CompressionException,
+            IllegalArgumentException,
+            ExpiredJwtException,
+            SignatureException {
+        try {
+            Jws<Claims> claimsJws = this.jwtParser.parseClaimsJws(token);
+            return claimsJws;
+        } catch (UnsupportedJwtException | MalformedJwtException | CompressionException | IllegalArgumentException e) {
+            throw UnauthenticatedException.MALFORMED_TOKEN;
+        } catch (ExpiredJwtException e) {
+            throw  UnauthenticatedException.EXPIRED_TOKEN;
+        } catch (io.jsonwebtoken.security.SignatureException e) {
+            throw UnauthenticatedException.INVALID_TOKEN;
+        }
     }
 
     public void injectAuthenticationTokenCookie(HttpServletResponse response, String cookieValue) {
@@ -135,15 +148,15 @@ public class AuthenticationService {
         tokenCookie.setHttpOnly(true);
         tokenCookie.setPath("/");
         tokenCookie.setSecure(false);
-        tokenCookie.setDomain("mylocal.tickers.local");
+        tokenCookie.setDomain(this.HOST_DOMAIN);
         response.addCookie(tokenCookie);
     }
 
     public void injectRefreshTokenCookie(HttpServletResponse response, String cookieValue) {
         Cookie refreshCookie = new Cookie("tickers.refresh", cookieValue);
         refreshCookie.setHttpOnly(true);
-        refreshCookie.setPath("/auth");
-        refreshCookie.setDomain("mylocal.tickers.local");
+        refreshCookie.setPath("/");
+        refreshCookie.setDomain(this.HOST_DOMAIN);
         refreshCookie.setSecure(false);
         response.addCookie(refreshCookie);
     }
@@ -164,15 +177,13 @@ public class AuthenticationService {
     }
 
     public String generateAccessToken(AuthenticationContext context, User user, UserSession session) {
-//        KeyWithId signingKey = jwtKeyProvider.getSigningKey("api.mlservice.local");
         return Jwts.builder()
                 .setSubject(user.getId().toString())
-//                .claim("tn", tenantContext.getTenantId())
                 .claim("ses", session.getId()) //Session ID - this can be revoked by the user to "kill" this session off or when the user logs out
                 .claim("typ", "access")
                 .claim("un", user.getUserName())
-                .setAudience("api.tickers.local")
-                .setIssuer("api.mlservice.local")
+                .setAudience(this.HOST_DOMAIN)
+                .setIssuer(this.ISSUER)
                 .setIssuedAt(new Date())
                 .setNotBefore(Optional.ofNullable(this.getSigningKey().getNotBefore()).map(zdt -> Date.from(zdt.toInstant())).orElse(null))
                 .setExpiration(Date.from(Instant.now().plus(10, ChronoUnit.MINUTES)))
@@ -183,20 +194,18 @@ public class AuthenticationService {
     }
 
     private KeyWithId getSigningKey() {
-        byte[] keyBytes = Decoders.BASE64.decode(this.secret);
-        KeyWithId key = new KeyWithId("api.mlservice.local", Keys.hmacShaKeyFor(keyBytes), null);
+        byte[] keyBytes = Decoders.BASE64.decode(this.SECRET);
+        KeyWithId key = new KeyWithId(this.ISSUER, Keys.hmacShaKeyFor(keyBytes), null);
         return key;
     }
 
     public String generateRefreshToken(AuthenticationContext context, User user, UserSession session) {
-//        KeyWithId signingKey = jwtKeyProvider.getSigningKey("api.mlservice.local");
         return Jwts.builder()
                 .claim("typ", "refresh")
-//                .claim("tn", tenantContext.getTenantId())
                 .setSubject(user.getId().toString())
                 .claim("ses", session.getId()) //Session ID - this can be revoked by the user to "kill" this session off or when the user logs out
-                .setAudience("api.tickers.local")
-                .setIssuer("api.mlservice.local")
+                .setAudience(this.HOST_DOMAIN)
+                .setIssuer(this.ISSUER)
                 .setIssuedAt(new Date())
                 .setNotBefore(Optional.ofNullable(this.getSigningKey().getNotBefore()).map(zdt -> Date.from(zdt.toInstant())).orElse(null))
                 .setExpiration(Date.from(session.getExpiresAt().toInstant()))
@@ -204,6 +213,21 @@ public class AuthenticationService {
                 .signWith(this.getSigningKey().getKey())
                 .compressWith(CompressionCodecs.GZIP)
                 .compact();
+    }
+
+    public void removeAllCookieHaveDomain(HttpServletRequest request, HttpServletResponse response) {
+        if (request.getCookies() != null) {
+            List<Cookie> cookieList = List.of(request.getCookies());
+
+            List<Cookie> tokenFiltered = cookieList.stream().filter(x -> x.getName().contains("tickers")).collect(Collectors.toList());
+
+            tokenFiltered.forEach(selectCookie -> {
+                selectCookie.setMaxAge(0);
+                selectCookie.setPath("/");
+                selectCookie.setDomain(this.HOST_DOMAIN);
+                response.addCookie(selectCookie);
+            });
+        }
     }
 
     public void logout(HttpServletRequest request, HttpServletResponse response) {
@@ -216,21 +240,27 @@ public class AuthenticationService {
 
         Cookie refreshTokenCookie = Arrays.stream(request.getCookies()).filter(c -> c.getName().equals("tickers.refresh"))
                 .findFirst().orElseThrow(() -> UnauthenticatedException.UNAUTHENTICATED);
+
+        Cookie tokenCookie = Arrays.stream(request.getCookies()).filter(c -> c.getName().equals("tickers.token"))
+                .findFirst().orElseThrow(() -> UnauthenticatedException.UNAUTHENTICATED);
         if (refreshTokenCookie.getValue().isBlank())
             throw UnauthenticatedException.UNAUTHENTICATED;
 
-        this.injectAuthenticationTokenCookie(response,  "");
-        this.injectRefreshTokenCookie(response, "");
+        if (tokenCookie.getValue().isBlank())
+            throw UnauthenticatedException.UNAUTHENTICATED;
 
         Jws<Claims> refreshTokenParsed;
         try {
-            refreshTokenParsed = this.jwtParser.parseClaimsJws(refreshTokenCookie.getValue());
+            refreshTokenParsed = this.parseJwt(refreshTokenCookie.getValue());
+            this.removeAllCookieHaveDomain(request, response);
             if (!Optional.ofNullable(refreshTokenParsed.getBody().get("typ", String.class)).orElseThrow(() -> new JwtException("Missing required claim: typ")).equals("refresh")) {
                 throw new JwtException("Invalid value for claim: typ");
             }
             UUID sessionId = UUID.fromString(Optional.ofNullable(refreshTokenParsed.getBody().get("ses", String.class)).orElseThrow(() -> new JwtException("Missing required claim: ses")));
             this.userSessionRepository.deleteById(sessionId);
             logger.info("Logged out session: {}", sessionId);
+
+
         } catch (JwtException e) {
             logger.error("There was an exception parsing the refresh token during logout. The user will be logged out but there session will not be removed", e);
             return;
